+++ docs/src/routes/+layout.svelte
@@ -791,6 +791,9 @@
 						<SidebarItem href="/splitter" labelText="Splitter" active={$page.url.pathname === '/splitter'}>
 							{#snippet icon()}•{/snippet}
 						</SidebarItem>
+						<SidebarItem href="/container-breakpoint" labelText="Responsive Fit" active={$page.url.pathname === '/container-breakpoint'}>
+							{#snippet icon()}•{/snippet}
+						</SidebarItem>
 					{/snippet}
 				</SidebarItem>